+++ src/compas_model/model/model.py
@@ -7,19 +7,39 @@ from compas.data import Data
 
 
 class Model(Data):
-    """model represents:\n
-    a dictionary of elements, where the key is the element.guid\n
-    a tree to represent the assembly hierarchy\n
-    a graph to represent the connectivity of the elements\n
+    """The Model data-structure represents:
+
+    a) flat collection of elements - dict{``uuid.uuid4()``, :class:`compas_model.elements.Element`}
+
+    b) hierarchical relationships between elements - :class:`compas.datastructures.Tree` (:class:`compas_model.model.ElementNode` or :class:`compas_model.model.GroupNode`)
+
+    c) abstract linkages (connections between elements and nodes) - :class:`compas.datastructures.Graph` (str(``uuid.uuid4()``), str(``uuid.uuid4()``))
 
     Parameters
     ----------
     name : str, optional
         A name or identifier for the model.
-    elements : list[:class:`compas_model.element.Element`], optional
+    elements : list[:class:`compas_model.elements.Element`], optional
         A list of elements to be added to the model.
     copy_elements : bool, optional
-        If True, the elements are copied before adding to the model.
+        If True, the :class:`compas_model.elements.Element` are copied before adding to the model.
+
+    Attributes
+    ----------
+    name : str
+        Name of the model.
+    elements : dict{``uuid.uuid4()``, :class:`compas_model.elements.Element`}
+        A dictionary of elements.
+    hierarchy : :class:`compas.datastructures.Tree`
+        A tree data-structure that stores a tree of elements and group nodes.
+    interactions : :class:`compas.datastructures.Graph`
+        A graph data-structure that stores interactions between elements.
+    number_of_elements : int
+        A total count in the :class:`compas_model.elements.Element` dictionary.
+    number_of_nodes : int
+        A total count of all :class:`compas.datastructures.TreeNode` in the hierarchy.
+    number_of_edges : int
+        A total count of all edges in the :class:`compas.datastructures.Graph`.
 
     """
 
@@ -27,15 +47,16 @@ class Model(Data):
         super(Model, self).__init__()
 
         # --------------------------------------------------------------------------
-        # initialize the main properties of the model
-        # a flat collection of elements - dict{GUID, Element}
-        # a hierarchical relationships between elements
-        # an abstract linkage or connection between elements and nodes
+        # Initialize the main properties of the model:
+        # a) flat collection of elements - dict{guid, Element}
+        # b) hierarchical relationships between elements - Tree(ElementNode or GroupNode)
+        # c) abstract linkages (connection between elements and nodes) - Graph(str(guid), str(guid))
         # --------------------------------------------------------------------------
         self._name = name  # the name of the model
         self._elements = OrderedDict()
         self._hierarchy = ElementTree(model=self, name=name)
         self._interactions = Graph(name=name)
+
         # --------------------------------------------------------------------------
         # if the user write Model(elements=[...])
         # then add the elements as ElementNode objects to the model
@@ -45,9 +66,9 @@ class Model(Data):
     # ==========================================================================
     # Serialization
     # ==========================================================================
+
     @property
     def data(self):
-
         return {
             "name": self._name,
             "elements": self._elements,
@@ -67,6 +88,7 @@ class Model(Data):
     # ==========================================================================
     # Attributes
     # ==========================================================================
+
     @property
     def name(self):
         return self._name
@@ -83,32 +105,12 @@ class Model(Data):
     def interactions(self):
         return list(self._interactions.edges())
 
-    # ==========================================================================
-    # Statistics
-    # ==========================================================================
     @property
     def number_of_elements(self):
-        """Get the number of elements in the model.
-
-        Returns
-        -------
-        int
-            The total number of elements in the model.
-
-        """
         return len(list(self.elements))
 
     @property
     def number_of_nodes(self):
-        """Count the total number of children in the tree hierarchy.
-
-        Returns
-        -------
-        int
-            The total number of child nodes in the tree hierarchy.
-
-        """
-
         def _count(node):
             count = 0
             if node.children is None:
@@ -122,59 +124,11 @@ class Model(Data):
 
     @property
     def number_of_edges(self):
-        """Get the number of edges in the model's interactions.
-
-        Returns
-        -------
-        int
-            The total number of edges in the interactions graph of the model.
-
-        """
         return self._interactions.number_of_edges()
 
     # ==========================================================================
     # Printing
     # ==========================================================================
-    def print_elements(self):
-        """Print all elements in the model.
-
-        This method prints all elements in the model to the console.
-
-        """
-        print(
-            "================================== {} ===================================".format(
-                self.interactions.name
-            )
-        )
-        graph_nodes = list(self._interactions.nodes())
-        for idx, e in enumerate(self._elements):
-            print(
-                "element_guid: "
-                + str(self._elements[e].guid)
-                + " graph_node: "
-                + str(graph_nodes[idx])
-            )
-
-    def print_interactions(self):
-        """Print all interactions between elements.
-        This method prints all interactions between elements in the model to the console.
-
-        """
-        print(
-            "================================== {} ===================================".format(
-                self._interactions.name
-            )
-        )
-        edges = list(self._interactions.edges())
-        for i in range(len(edges)):
-            a = edges[i][0]
-            b = edges[i][1]
-            print(
-                "print_interactions ",
-                str(self._elements[a].guid),
-                " ",
-                str(self._elements[b].guid),
-            )
 
     def __repr__(self):
         return (
@@ -193,10 +147,11 @@ class Model(Data):
         return self.__repr__()
 
     def print(self):
-        """Print the spatial hierarchy of the tree for debugging and visualization.
+        """Print the spatial strucutre of the :class:`compas_model.model.ElementTree`,
+        also total number of :class:`compas_model.elements.Element`, :class:`compas.datastructures.Graph` nodes and edges.
 
         This method prints information about the tree's spatial hierarchy, including nodes, elements,
-        parent-child relationships, and other relevant details.
+        parent-child relationships.
 
         """
         # ------------------------------------------------------------------
@@ -245,26 +200,58 @@ class Model(Data):
             print(" " * 4 + str(edge[0]) + " " + str(edge[1]))
         print("\u2500" * 100)
 
+    def print_elements(self):
+        """Print all :class:`compas_model.elements.Element` in the model."""
+        print(
+            "================================== {} ===================================".format(
+                self.interactions.name
+            )
+        )
+        graph_nodes = list(self._interactions.nodes())
+        for idx, e in enumerate(self._elements):
+            print(
+                "element_guid: "
+                + str(self._elements[e].guid)
+                + " graph_node: "
+                + str(graph_nodes[idx])
+            )
+
+    def print_interactions(self):
+        """Print all :class:`compas.datastructures.Graph` nodes and edges."""
+        print(
+            "================================== {} ===================================".format(
+                self._interactions.name
+            )
+        )
+        edges = list(self._interactions.edges())
+        for i in range(len(edges)):
+            a = edges[i][0]
+            b = edges[i][1]
+            print(
+                "print_interactions ",
+                str(self._elements[a].guid),
+                " ",
+                str(self._elements[b].guid),
+            )
+
     # ==========================================================================
     # Behavior - Hierarchy
     # ==========================================================================
 
     def add_elements(self, elements=[], copy_elements=False):
-        """Adds elements to the model.
-
-        This method allows you to add elements to the model.
+        """Multiple calls to the :class:`compas_model.model.Model.add_element` method.
 
         Parameters
         ----------
         elements : list[:class:`compas_model.elements.Element`], optional
-            A list of elements to be added to the model.
+            A list of elements.
         copy_elements : bool, optional
             If True, the elements are copied before adding to the model.
 
         Returns
         -------
-        list : guid
-            A list of identifiers for the elements.
+        list : ``uuid.uuid4()``
+            A list of identifiers of the elements.
 
         """
         guids = []
@@ -277,9 +264,7 @@ class Model(Data):
         return guids
 
     def add_element(self, name=None, element=None, attributes=None, copy_element=False):
-        """Adds an element to the model.
-
-        This method allows you to add an element to the model.
+        """Add a :class:`compas_model.model.ElementNode` that represents a leaf with a property of an :class:`compas_model.elements.Element`.
 
         Parameters
         ----------
@@ -306,7 +291,7 @@ class Model(Data):
         )
 
     def add_group(self, name=None, geometry=None, attributes=None):
-        """Adds a group to the model.
+        """Add a :class:`compas_model.model.GroupNode` that represent a group.
 
         Parameters
         ----------
@@ -332,40 +317,21 @@ class Model(Data):
     # ==========================================================================
     # Behavior - Interactions
     # ==========================================================================
-    # def add_interaction_node(self, name):
-    #     """Adds an interaction node to the model.
-
-    #     This method allows you to add an interaction node to the model.
-
-    #     Parameters
-    #     ----------
-    #     name : str
-    #         A name or identifier for the node.
-
-    #     Returns
-    #     -------
-    #     hashable
-    #         The identifier of the node.
-
-    #     """
-
-    #     return self._interactions.add_node(name)
 
     def add_interaction(self, element0, element1, geometry=None, weight=1):
-        """Adds an interaction between two elements in the model.
-
-        This method allows you to establish an interaction between two elements within the model.
+        """Add edges as a pair of :class:`compas_model.elements.Element` str(``uuid.uuid4()`` to the :class:`compas.datastructures.Graph`.
+        The :class:`compas_model.model.Model.interactions` already contains all the previously added elements identifiers.
 
         Parameters
         ----------
-        element0 : :class:`compas_model.elements.Element` or :class:`compas_model.model.ElementNode
+        element0 : :class:`compas_model.elements.Element` or :class:`compas_model.model.ElementNode`
             The first element involved in the interaction.
-        element1 : :class:`compas_model.elements.Element` or :class:`compas_model.model.ElementNode
+        element1 : :class:`compas_model.elements.Element` or :class:`compas_model.model.ElementNode`
             The second element involved in the interaction.
 
         Returns
         -------
-        tuple[str, str]
+        tuple[str(``uuid.uuid4()``), str(``uuid.uuid4()``)]
             The identifier of the edge.
 
         """
@@ -388,20 +354,22 @@ class Model(Data):
                 str(e0.guid), str(e1.guid), attribute_dict
             )
         else:
-            raise ValueError("The node does not exist.")
+            raise ValueError("The Node does not exist.")
 
     # ==========================================================================
     # Copy
     # ==========================================================================
+
     def copy(self):
-        """copy the model and duplicate all elements, tree nodes and graph nodes"""
+        """Duplicate the :class:`compas_model.model.Model` properties: dict{``uuid.uuid4()``,
+        :class:`compas_model.elements.Element`}, :class:`compas.datastructures.Tree` and :class:`compas.datastructures.Graph`."""
         # --------------------------------------------------------------------------
-        # create the empty model
+        # Create the empty model.
         # --------------------------------------------------------------------------
         copy = Model(name=self.name)
 
         # --------------------------------------------------------------------------
-        # copy the hierarchy
+        # Copy the hierarchy.
         # --------------------------------------------------------------------------
         dict_old_guid_and_new_element = {}
 
@@ -409,31 +377,31 @@ class Model(Data):
             for child in current_node.children:
                 last_group_node = None
                 # --------------------------------------------------------------------------
-                # copy the elements
+                # Copy the elements.
                 # --------------------------------------------------------------------------
                 if isinstance(child, ElementNode):
-                    # copy the element
+                    # Copy the Element.
                     name = child.name
                     element = child.element.copy()
-                    # add the element to the dictionary
+                    # Add the Element to the Dictionary.
                     copy._elements[element.key] = element
-                    # add the element to the graph
+                    # Add the Element to the Graph.
                     copy._interactions.add_node(element.key)
-                    # add the element to the parent
+                    # Add the Element to the Parent Node.
                     copy_node.add_element(name=name, element=element)
-                    # add the element to the dictionary
+                    # Add the Element to the Model Dictionary.
                     dict_old_guid_and_new_element[child.element.key] = element
                 # --------------------------------------------------------------------------
-                # copy the groups
+                # Copy the groups.
                 # --------------------------------------------------------------------------
                 elif isinstance(child, GroupNode):
-                    # copy the group
+                    # Copy the group.
                     name = child.name
                     geometry = None if child.geometry is None else child.geometry.copy()
-                    # add the group to the parent
+                    # Add the group to the Parent Node.
                     last_group_node = copy_node.add_group(name=name, geometry=geometry)
                 # --------------------------------------------------------------------------
-                # recursively copy the tree
+                # Recursively copy the Tree.
                 # --------------------------------------------------------------------------
                 if isinstance(child, GroupNode):
                     copy_hierarchy(child, last_group_node)
@@ -441,7 +409,7 @@ class Model(Data):
         copy_hierarchy(self._hierarchy.root, copy._hierarchy.root)
 
         # --------------------------------------------------------------------------
-        # copy the interactions, nodes should be added previously
+        # Copy the interactions, Nodes should be added previously.
         # --------------------------------------------------------------------------
         for edge in self._interactions.edges():
             node0 = dict_old_guid_and_new_element[edge[0]]